update tour startdate

+++ src/java/controller/DeletePlayerServlet.java
@@ -5,6 +5,7 @@
 package controller;
 
 import dao.PlayerDAO;
+import dao.TeamDAO;
 import java.io.IOException;
 import java.io.PrintWriter;
 import java.util.logging.Level;
@@ -13,6 +14,7 @@ import javax.servlet.ServletException;
 import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
+import model.Team;
 
 /**
  *
@@ -36,13 +38,19 @@ public class DeletePlayerServlet extends HttpServlet {
         int player_id = Integer.parseInt(request.getParameter("player_id"));
         int user_id = Integer.parseInt(request.getParameter("user_id"));
 
+        TeamDAO teamDAO = new TeamDAO();
+        Team team = teamDAO.getTeamByID(team_id);
+
         try {
             PlayerDAO playerDAO = new PlayerDAO();
             if (playerDAO.IsTeamManager(user_id, team_id)) {
-                playerDAO.deletePlayer(player_id);
+                if (playerDAO.isPlayerCanBeDelete(player_id)) {
+                    playerDAO.deletePlayer(player_id);
+                }
             }
-            request.getRequestDispatcher("manage/teamDetail.jsp").forward(request, response);
-
+            request.setAttribute("team_id", team_id);
+            request.setAttribute("coach", team.getCoach());
+            response.sendRedirect("teamDetail?team_id="+team_id+"&coach="+team.getCoach());
         } catch (Exception ex) {
             Logger.getLogger(ParticipantListServlet.class.getName()).log(Level.SEVERE, null, ex);
         }
